Add backface culling for perspective projection
This works well and as expected for perspective projection.

If I switch it to orthographic projection the faces are culled before
they should be. Switch is by removing the perspective divide (x/z,
y/z) and lowering the fov factor.

I'm thinking the matrices will fix this so I'm going to carry on.